add samples

+++ gpu/texture.cpp
@@ -39,15 +39,15 @@ math::vec4f Texture::getColor(float u, float v) {
         resultColor = mBuffer[position];
     }
     else if (mFilter == TEXTURE_FILTER_LINEAR) {
-        float x = u * (mWidth - 1);
-        float y = v * (mHeight - 1);
+        float x = u * static_cast<float>(mWidth - 1);
+        float y = v * static_cast<float>(mHeight - 1);
 
         int left = std::floor(x);  //std::floor向下取整
         int right = std::ceil(x);  //std::ceil向上取整
         int bottom = std::floor(y);
         int top = std::ceil(y);
 
-        //对上下差值，得到左右
+        //对上下插值，得到左右
         float yScale = 0.0f;
         if (top == bottom) {
             yScale = 1.0f;
@@ -64,7 +64,7 @@ math::vec4f Texture::getColor(float u, float v) {
         RGBA leftColor = math::lerp(mBuffer[positionLeftBottom], mBuffer[positionLeftTop], yScale);
         RGBA rightColor = math::lerp(mBuffer[positionRightBottom], mBuffer[positionRightTop], yScale);
 
-        //对左右差值，得到结果
+        //对左右插值，得到结果
         float xScale = 0.0f;
         if (right == left) {
             xScale = 1.0f;
